+++ jest.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  testPathIgnorePatterns: ['/node_modules/', '/.next/'],
+  setupFilesAfterEnv: ['<rootDir>/react/tests/setupTests.ts'],
+  transform: {
+    '^.+\\.(js|jsx|ts|tsx)$': '<rootDir>/node_modules/babel-jest',
+  },
+  testEnvironment: 'jsdom',
+  moduleDirectories: ['node_modules'],
+}