+++ react/__mocks__/vtex.styleguide.tsx
@@ -1,0 +1,6 @@
+export { default as Alert } from '@vtex/styleguide/lib/Alert'
+export { default as Collapsible } from '@vtex/styleguide/lib/Collapsible'
+export { default as InputButton } from '@vtex/styleguide/lib/InputButton'
+export { default as IconInfo } from '@vtex/styleguide/lib/icon/Info'
+export { default as Tooltip } from '@vtex/styleguide/lib/Tooltip'
+export { default as Divider } from '@vtex/styleguide/lib/Divider'