+++ CHANGELOG.md
@@ -20,15 +20,22 @@ All notable changes to this project will be documented in this file.
 - Tests
 
 ### Changed 
-- Correction of repost bugs by ACCT
+- Correction of reported bugs by ACCT
   - Change style of wallet to mobile
   - Change the input "Enter the value" to only accept numbers
-  - Change the format orf money change the formatting of the money to follow the standard (R$ 1.000,00)
+  - Change the formatting of the money to follow the standard (R$ 1.000,00)
   - Change the style of Input Buttons
   - Change the width of divider
   - Add success or error mensages when click to button "Copy Code"
   - Change message of tooltip
   - Add line break for mobiles in the balance data area 
+  - Change InputButton to InputCurrency and Button
+  - Cleaning field after success in generating redemption value
+  - Change Alert to Toast
+  - Add timer on Toast
+  - Change InputCurrency to IntlCurrencyInput
+  - Add red border to sinalize a error
+
 
 ### Added
 - Update tests