+++ CHANGELOG.md
@@ -37,4 +37,5 @@ All notable changes to this project will be documented in this file.
   - Add red border to sinalize a error
 
 
-
+### Added
+- Update tests